ScheduleTreeElemBand::drop: change type of position and number to size_t
This is a more natural type for a position and a number.

+++ tc/core/polyhedral/schedule_tree_elem.cc
@@ -151,9 +151,9 @@ size_t ScheduleTreeElemBand::nOuterCoincident() const {
   return i;
 }
 
-void ScheduleTreeElemBand::drop(int pos, int n) {
-  CHECK_LE(0, n) << "range out of bounds";
-  CHECK_LE(0, pos) << "range  out of bounds";
+void ScheduleTreeElemBand::drop(size_t pos, size_t n) {
+  CHECK_LE(0u, n) << "range out of bounds";
+  CHECK_LE(0u, pos) << "range  out of bounds";
   CHECK_GE(nMember(), pos + n) << "range out of bounds";
   int nBegin = nMember();
